various
refactoring, removed abandoned lines and classes, included more of the
dtd into the xsd
