+++ express2xml_descriptionsLocal/.settings/org.eclipse.wst.common.project.facet.core.prefs.xml
@@ -1,7 +1,7 @@
 <root>
   <facet id="jpt.jaxb">
     <node name="classgen">
-      <attribute name="package" value="main.java.com.boeing.stepdev.util.express2xml.jaxb"/>
+      <attribute name="package" value="stepmod.express2xml.descriptions.util.jaxb"/>
     </node>
     <node name="libprov">
       <attribute name="provider-id" value="jaxb-user-library-provider"/>
@@ -11,7 +11,7 @@
     </node>
     <node name="schemas">
       <node name="schema-1">
-        <attribute name="location" value="platform:/resource/express2xml_descriptions/src/stepdev/express2xml/descriptions/resources/description.xsd"/>
+        <attribute name="location" value="platform:/resource/express2xml_descriptions/src/stepmod/express2xml/descriptions/resources/description.xsd"/>
       </node>
     </node>
   </facet>
