--- express2xml_descriptionsLocal/src/stepdev/express2xml/descriptions/resources/description.xsd
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<schema xmlns:ns="thisSchema"
-xmlns="http://www.w3.org/2001/XMLSchema"
-targetNamespace="thisSchema"
-elementFormDefault="qualified"
-attributeFormDefault="qualified">
-
-<!-- simple types -->
-<simpleType name="linkendType">
-	<restriction base="string"/>
-</simpleType>
-
-<!-- complex types -->
-<complexType name="express_refType">
-	<attribute ref="ns:linkend" use="required"/>
-</complexType>
-
-<complexType name="ext_descriptionType" mixed="true">
-	<sequence>
-		<element ref="ns:express_ref"/>
-	</sequence>
-	<attribute ref="ns:linkend" use="required"/>
-</complexType>
-
-<complexType name="ext_descriptionsType">
-	<sequence>
-		<element ref="ns:ext_description" minOccurs="1" maxOccurs="unbounded"/>
-	</sequence>
-</complexType>
-
-<!-- elements -->
-<element name="express_ref" type="ns:express_refType"/>
-<element name="ext_description" type="ns:ext_descriptionType"/>
-
-<!-- attributes -->
-<attribute name="linkend" type="ns:linkendType"/>
-
-<!-- main -->
-<element name="ext_descriptions" type="ns:ext_descriptionsType"/>
-
-</schema>
