--- express2xml_descriptionsLocal/src/stepdev/express2xml/descriptions/util/jaxb/Ext_descriptionsType.java
@@ -1,76 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, vhudson-jaxb-ri-2.2-147 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.04.17 at 04:58:56 PM PDT 
-//
-
-
-package stepdev.express2xml.descriptions.util.jaxb;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for ext_descriptionsType complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="ext_descriptionsType">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element ref="{thisSchema}ext_description" maxOccurs="unbounded"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "ext_descriptionsType", propOrder = {
-    "extDescription"
-})
-public class Ext_descriptionsType {
-
-    @XmlElement(name = "ext_description", required = true)
-    protected List<ExtDescriptionType> extDescription;
-
-    /**
-     * Gets the value of the extDescription property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the extDescription property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getExtDescription().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link ExtDescriptionType }
-     * 
-     * 
-     */
-    public List<ExtDescriptionType> getExtDescription() {
-        if (extDescription == null) {
-            extDescription = new ArrayList<ExtDescriptionType>();
-        }
-        return this.extDescription;
-    }
-
-}
