+++ express2xml_descriptionsLocal/src/stepmod/express2xml/descriptions/util/jaxb/ExtDescriptionType.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, vhudson-jaxb-ri-2.2-147 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.04.17 at 04:58:56 PM PDT 
+// Generated on: 2018.04.24 at 07:52:35 PM PDT 
 //
 
 
-package stepdev.express2xml.descriptions.util.jaxb;
+package stepmod.express2xml.descriptions.util.jaxb;
 
 import java.io.Serializable;
 import java.util.ArrayList;
