+++ express2xml_descriptionsLocal/src/stepmod/express2xml/descriptions/util/jaxb/ObjectFactory.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, vhudson-jaxb-ri-2.2-147 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.04.17 at 04:58:56 PM PDT 
+// Generated on: 2018.04.24 at 07:52:35 PM PDT 
 //
 
 
-package stepdev.express2xml.descriptions.util.jaxb;
+package stepmod.express2xml.descriptions.util.jaxb;
 
 import javax.xml.bind.JAXBElement;
 import javax.xml.bind.annotation.XmlElementDecl;
@@ -17,7 +17,7 @@ import javax.xml.namespace.QName;
 /**
  * This object contains factory methods for each 
  * Java content interface and Java element interface 
- * generated in the main.java.com.boeing.stepdev.util.express2xml.jaxb package. 
+ * generated in the stepmod.express2xml.descriptions.util.jaxb package. 
  * <p>An ObjectFactory allows you to programatically 
  * construct new instances of the Java representation 
  * for XML content. The Java representation of XML 
@@ -31,15 +31,23 @@ import javax.xml.namespace.QName;
 @XmlRegistry
 public class ObjectFactory {
 
-    private final static QName _Ext_descriptions_QNAME = new QName("thisSchema", "ext_descriptions");
+    private final static QName _ExtDescriptions_QNAME = new QName("thisSchema", "ext_descriptions");
     private final static QName _ExpressRef_QNAME = new QName("thisSchema", "express_ref");
     private final static QName _ExtDescription_QNAME = new QName("thisSchema", "ext_description");
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: main.java.com.boeing.stepdev.util.express2xml.jaxb
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: stepmod.express2xml.descriptions.util.jaxb
      * 
      */
     public ObjectFactory() {
+    }
+
+    /**
+     * Create an instance of {@link ExtDescriptionsType }
+     * 
+     */
+    public ExtDescriptionsType createExtDescriptionsType() {
+        return new ExtDescriptionsType();
     }
 
     /**
@@ -59,20 +67,12 @@ public class ObjectFactory {
     }
 
     /**
-     * Create an instance of {@link Ext_descriptionsType }
-     * 
-     */
-    public Ext_descriptionsType createExt_descriptionsType() {
-        return new Ext_descriptionsType();
-    }
-
-    /**
-     * Create an instance of {@link JAXBElement }{@code <}{@link Ext_descriptionsType }{@code >}}
+     * Create an instance of {@link JAXBElement }{@code <}{@link ExtDescriptionsType }{@code >}}
      * 
      */
     @XmlElementDecl(namespace = "thisSchema", name = "ext_descriptions")
-    public JAXBElement<Ext_descriptionsType> createExt_descriptions(Ext_descriptionsType value) {
-        return new JAXBElement<Ext_descriptionsType>(_Ext_descriptions_QNAME, Ext_descriptionsType.class, null, value);
+    public JAXBElement<ExtDescriptionsType> createExtDescriptions(ExtDescriptionsType value) {
+        return new JAXBElement<ExtDescriptionsType>(_ExtDescriptions_QNAME, ExtDescriptionsType.class, null, value);
     }
 
     /**
